requirements.txt
Mezzanine
psycopg2
django-tinymce
django_pandas
numpy
quandl
psycopg2
scrapy
Cython
twisted
